# backend/openapi.yaml
components:
  schemas:
    APIArea:
      description: エリア情報のAPIモデル
      properties:
        id:
          title: Id
          type: integer
        name:
          title: Name
          type: string
      required:
        - id
        - name
      title: APIArea
      type: object
    APICell:
      description: セル情報のAPIモデル
      properties:
        geo_hash:
          title: Geo Hash
          type: string
        id:
          title: Id
          type: integer
        location:
          maxItems: 2
          minItems: 2
          prefixItems:
            - type: number
            - type: number
          title: Location
          type: array
      required:
        - id
        - geo_hash
        - location
      title: APICell
      type: object
    APIPost-Input:
      description: 投稿情報のAPIモデル（ユーザー投稿・LLM返信）
      properties:
        area:
          anyOf:
            - $ref: "#/components/schemas/APIArea"
            - type: "null"
        cell:
          anyOf:
            - $ref: "#/components/schemas/APICell"
            - type: "null"
        content:
          title: Content
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        location:
          maxItems: 2
          minItems: 2
          prefixItems:
            - type: number
            - type: number
          title: Location
          type: array
        replies:
          default: []
          items:
            $ref: "#/components/schemas/APIPost-Input"
          title: Replies
          type: array
        user_uuid:
          anyOf:
            - format: uuid
              type: string
            - type: "null"
          title: User Uuid
        uuid:
          format: uuid
          title: Uuid
          type: string
      required:
        - uuid
        - content
        - location
        - created_at
      title: APIPost
      type: object
    APIPost-Output:
      description: 投稿情報のAPIモデル（ユーザー投稿・LLM返信）
      properties:
        area:
          anyOf:
            - $ref: "#/components/schemas/APIArea"
            - type: "null"
        cell:
          anyOf:
            - $ref: "#/components/schemas/APICell"
            - type: "null"
        content:
          title: Content
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        location:
          maxItems: 2
          minItems: 2
          prefixItems:
            - type: number
            - type: number
          title: Location
          type: array
        replies:
          default: []
          items:
            $ref: "#/components/schemas/APIPost-Output"
          title: Replies
          type: array
        user_uuid:
          anyOf:
            - format: uuid
              type: string
            - type: "null"
          title: User Uuid
        uuid:
          format: uuid
          title: Uuid
          type: string
      required:
        - uuid
        - content
        - location
        - created_at
      title: APIPost
      type: object
    CreatePostRequest:
      description: 投稿作成エンドポイントのリクエスト
      properties:
        content:
          title: Content
          type: string
        lat:
          title: Lat
          type: number
        lon:
          title: Lon
          type: number
      required:
        - content
        - lat
        - lon
      title: CreatePostRequest
      type: object
    CreatePostResponse:
      description: 投稿作成エンドポイントのレスポンス
      properties:
        post:
          $ref: "#/components/schemas/APIPost-Output"
        similar_posts:
          items:
            $ref: "#/components/schemas/APIPost-Output"
          title: Similar Posts
          type: array
      required:
        - post
        - similar_posts
      title: CreatePostResponse
      type: object
    CurrentResponse:
      description: 現在位置エンドポイントのレスポンス
      properties:
        area:
          $ref: "#/components/schemas/APIArea"
        cell:
          $ref: "#/components/schemas/APICell"
      required:
        - area
        - cell
      title: CurrentResponse
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    PostsResponse:
      description: 投稿一覧エンドポイントのレスポンス
      properties:
        posts:
          items:
            $ref: "#/components/schemas/APIPost-Output"
          title: Posts
          type: array
      required:
        - posts
      title: PostsResponse
      type: object
    SignupResponse:
      description: サインアップエンドポイントのレスポンス
      properties:
        access_token:
          title: Access Token
          type: string
        refresh_token:
          title: Refresh Token
          type: string
        user_id:
          title: User Id
          type: string
      required:
        - access_token
        - refresh_token
        - user_id
      title: SignupResponse
      type: object
    UpdatePromptResponse:
      description: プロンプト更新エンドポイントのレスポンス
      properties:
        success:
          title: Success
          type: boolean
      required:
        - success
      title: UpdatePromptResponse
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
        - loc
        - msg
        - type
      title: ValidationError
      type: object
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.1.0
paths:
  /:
    get:
      description: ヘルスチェック
      operationId: health__get
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: Health
      tags:
        - health
  /areas:
    get:
      description: 全エリアの一覧を取得
      operationId: get_areas_areas_get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/APIArea"
                title: Response Get Areas Areas Get
                type: array
          description: Successful Response
      summary: Get Areas
      tags:
        - areas
  /batch/prompts/update:
    post:
      description: 全エリアのプロンプトをバッチ更新
      operationId: update_prompt_batch_prompts_update_post
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatePromptResponse"
          description: Successful Response
      summary: Update Prompt
      tags:
        - batch
  /current:
    get:
      description: 現在位置のエリアとセル情報を取得
      operationId: get_current_current_get
      parameters:
        - in: query
          name: lat
          required: true
          schema:
            title: Lat
            type: number
        - in: query
          name: lon
          required: true
          schema:
            title: Lon
            type: number
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Current
      tags:
        - current
  /posts:
    get:
      description: 現在位置の投稿一覧を取得
      operationId: get_posts_posts_get
      parameters:
        - in: query
          name: lat
          required: true
          schema:
            title: Lat
            type: number
        - in: query
          name: lon
          required: true
          schema:
            title: Lon
            type: number
        - in: header
          name: authorization
          required: true
          schema:
            title: Authorization
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Posts
      tags:
        - posts
    post:
      description: 新しい投稿を作成
      operationId: create_post_posts_post
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            title: Authorization
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePostResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Create Post
      tags:
        - posts
  /posts/me:
    get:
      description: 自分の投稿一覧を取得
      operationId: get_my_posts_posts_me_get
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            title: Authorization
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get My Posts
      tags:
        - posts
  /signup:
    post:
      description: 匿名ユーザーとしてサインアップ
      operationId: signup_signup_post
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupResponse"
          description: Successful Response
      summary: Signup
      tags:
        - auth